# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateTaskRequest:
    properties:
      description:
        example: описание задачи
        type: string
      due_date:
        example: "2025-11-20T13:37:27+00:00"
        type: string
      priority:
        example: low/medium/high
        type: string
      status:
        example: todo/in-progress/done
        type: string
      title:
        example: написать функцию
        minLength: 1
        type: string
    required:
    - title
    type: object
  dto.ErrorResponse:
    properties:
      error:
        example: Произошла ошибка
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  dto.SuccessResponse:
    properties:
      data: {}
      message:
        example: Операция выполнена успешно
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: Support Team
  description: API для управления задачами, проектами и комментариями
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: SaaS
  version: "1.0"
paths:
...
  /projects/{id}/tasks:
    get:
      consumes:
      - application/json
      description: Возвращает все задачи проекта. Требуются права просмотра проекта.
      parameters:
      - description: ID проекта
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить задачи проекта
      tags:
      - task
    post:
      consumes:
      - application/json
      description: Создаёт новую запись в таблице задач. Требуются права редактора
        проекта.
      parameters:
      - description: ID проекта
        in: path
        name: id
        required: true
        type: integer
      - description: Название, описание, статус, приоритет и дата завершения задачи
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать задачу
      tags:
      - task
...
  /tasks/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет запись в таблице задач. Требуются права редактора проекта.
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить
      tags:
      - task
    patch:
      consumes:
      - application/json
      description: Изменяет запись в таблице задач. Требуются права редактора проекта.
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      - description: Название, описание, статус, приоритет и дата завершения задачи
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Изменить задачу
      tags:
      - task
    post:
      consumes:
      - application/json
      description: Возвращает одну конкретную задачу. Требуются права просмотра проекта.
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить задачу
      tags:
      - task
...
securityDefinitions:
  BearerAuth:
    description: Введите "Bearer {token}" (без кавычек)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"